move slam/testhelper into existing internal/testhelper
