+++ integration_test.go
@@ -27,7 +27,6 @@ import (
 	vcConfig "github.com/viamrobotics/viam-cartographer/config"
 	"github.com/viamrobotics/viam-cartographer/dataprocess"
 	"github.com/viamrobotics/viam-cartographer/internal/testhelper"
-	vcTesthelper "github.com/viamrobotics/viam-cartographer/testhelper"
 )
 
 const (
@@ -90,7 +89,7 @@ func integrationtestHelperCartographer(t *testing.T, subAlgo viamcartographer.Su
 		t.Skip()
 	}
 
-	dataDir, err := vcTesthelper.CreateTempFolderArchitecture(logger)
+	dataDir, err := testhelper.CreateTempFolderArchitecture(logger)
 	test.That(t, err, test.ShouldBeNil)
 
 	prevNumFiles := 0
@@ -243,7 +242,7 @@ func integrationtestHelperCartographer(t *testing.T, subAlgo viamcartographer.Su
 	time.Sleep(time.Millisecond * cartoSleepMsec)
 
 	// Remove existing pointclouds, but leave maps and config (so we keep the lua files).
-	test.That(t, vcTesthelper.ResetFolder(dataDir+"/data"), test.ShouldBeNil)
+	test.That(t, testhelper.ResetFolder(dataDir+"/data"), test.ShouldBeNil)
 	prevNumFiles = 0
 
 	// Count the initial number of maps in the map directory (should equal 1)
@@ -311,7 +310,7 @@ func integrationtestHelperCartographer(t *testing.T, subAlgo viamcartographer.Su
 	testCartographerMap(t, svc)
 
 	// Remove maps so that testing is done on the map generated by the internal map
-	test.That(t, vcTesthelper.ResetFolder(dataDir+"/map"), test.ShouldBeNil)
+	test.That(t, testhelper.ResetFolder(dataDir+"/map"), test.ShouldBeNil)
 
 	testCartographerInternalState(t, svc, dataDir)
 
@@ -325,7 +324,7 @@ func integrationtestHelperCartographer(t *testing.T, subAlgo viamcartographer.Su
 	time.Sleep(time.Millisecond * cartoSleepMsec)
 
 	// Remove existing pointclouds, but leave maps and config (so we keep the lua files).
-	test.That(t, vcTesthelper.ResetFolder(dataDir+"/data"), test.ShouldBeNil)
+	test.That(t, testhelper.ResetFolder(dataDir+"/data"), test.ShouldBeNil)
 	prevNumFiles = 0
 
 	// Test online mode using the map generated in the offline test
